requirements/requirements.gpu.txt
numpy==1.15.0
scikit-learn[alldeps]==0.19.0
pandas==0.22.0
mxnet-cu90==1.3.0b20180820
